finish-test-cases

+++ src/services/lostItems.service.ts
@@ -20,8 +20,8 @@ export const getAllLostItemsService = async () => {
 
 export const getLostItemsByIdService = async (id: number) => {
 	const lostItems = await prisma.lost_items.findUnique({ where: { id } });
-	if (!lostItems) {
-		throw new Error(`${LOST_ITEM}_${DB_ERROR_CODES.NOT_FOUND}`);
+	if (!id || !lostItems) {
+		throw new Error(`${LOST_ITEM}_${id}_${DB_ERROR_CODES.NOT_FOUND}`);
 	}
 	return lostItems;
 }
@@ -37,16 +37,16 @@ export const createLostItemService = async (data: any) => {
 export const updateLostItemService = async (id: number, data: any) => {
 	if (!data?.date || !data?.room || !data?.description) {
 		throw new Error(`${LOST_ITEM}_${DB_ERROR_CODES.INVALID_DATA}`);
-	} else if (!(await prisma.lost_items.findUnique({ where: { id } }))) {
-		throw new Error(`${LOST_ITEM}_${DB_ERROR_CODES.NOT_FOUND}`);
+	} else if (!id || !(await prisma.lost_items.findUnique({ where: { id } }))) {
+		throw new Error(`${LOST_ITEM}_${id}_${DB_ERROR_CODES.NOT_FOUND}`);
 	}
 	const updatedLostItem = await prisma.lost_items.update({ where: { id }, data });
 	return updatedLostItem;
 };
 
 export const deleteLostItemService = async (id: number) => {
-	if (!(await prisma.lost_items.findUnique({ where: { id } }))) {
-		throw new Error(`${LOST_ITEM}_${DB_ERROR_CODES.NOT_FOUND}`);
+	if (!id || !(await prisma.lost_items.findUnique({ where: { id } }))) {
+		throw new Error(`${LOST_ITEM}_${id}_${DB_ERROR_CODES.NOT_FOUND}`);
 	}
 	const deletedLostItem = await prisma.lost_items.delete({ where: { id } });
 	return deletedLostItem;
